requirements.txt
django==1.7
django-rest-auth==0.3.4
djangorestframework==3.3.2
djangocms-admin-style==0.2.5
django-rest-swagger==0.2.9
python-social-auth==0.2.12

pycrypto==2.6.1

psycopg2==2.6
dj-database-url

uwsgi==2.0.10

virtualenv==12.0.7
pyjade
django-pipeline<1.6
jsmin
watchdog
supervisor-stdout

